Currently slow queries are occurred in `PrecomputeFeedService`.
Honestly, I can't explain this problem and solution in detail. I'm not familiar with PostgreSQL.

- This problem often occurred a day.
- I confirmed amount of account's followers. Account has 10 ~ 1000> followers.
- It seems that the problem is caused by the query plan.
  - I don't know why, but this PR way I solved it.

I want to ask to for your advise about this PR. 😕  🙏 

---

**before**

`EXPLAIN ANALYZE SELECT "statuses".* FROM "statuses" WHERE "statuses"."account_id" IN (1, 2, 3...) ORDER BY id desc LIMIT 100;`

```
Limit  (cost=0.43..2113.78 rows=100 width=316) (actual time=978.410..12230.796 rows=100 loops=1)
 ->  Index Scan Backward using statuses_pkey on statuses  (cost=0.43..16140231.67 rows=763730 width=316) (actual time=978.409..12230.770 rows=100 loops=1)
  Filter: (account_id = ANY ('{1, 2, 3...}'::integer[]))
       Rows Removed by Filter: 1701437
Planning time: 2.141 ms
Execution time: 12230.841 ms
(6 rows)
```

**after**

Add `account_id desc` to `ORDER BY` section.

`EXPLAIN ANALYZE SELECT "statuses".* FROM "statuses" WHERE "statuses"."account_id" IN (1, 2, 3...) ORDER BY id desc, accont_id desc LIMIT 100;`

```
Limit  (cost=1273363.52..1273363.77 rows=100 width=316) (actual time=406.579..406.605 rows=100 loops=1)
 ->  Sort  (cost=1273363.52..1275272.85 rows=763730 width=316) (actual time=406.579..406.597 rows=100 loops=1)
       Sort Key: id DESC, account_id DESC
       Sort Method: top-N heapsort  Memory: 65kB
       ->  Index Scan using index_statuses_on_account_id on statuses  (cost=0.43..1244174.31 rows=763730 width=316) (actual time=0.058..388.025 rows=177439 loops=1)
        Index Cond: (account_id = ANY ('{1, 2, 3...}'::integer[]))
Planning time: 2.140 ms
Execution time: 406.672 ms
(8 rows)
```