@S-YOU 2017-05-11
May be this is the best with heapsort, and you have mixed index, which used less memory, stop when limit is reached. 

https://madusudanan.com/blog/all-you-need-to-know-about-sorting-in-postgres/#Heapsort

And first one removed 1701437 rows after it loaded into memory. (Rows Removed by Filter: 1701437)