@alpaca-tc 2017-05-11
Sorry, unfortunately it seems that this PR contains the side effect.

- (too slow-> fast) If account's followers don't toots much: sorting records are faster than primary key scan. This is nice for me.
- side effect - (fast -> slow) If account has many followers and followers frequently toots: sorting records are slower than primary key scan. 😕 

I will see a slow log of pawoo.net for a while. 🙏