@Gargron 2017-05-11
If it's any consolation, this query should be fairly rare and only done in background processing.