@alpaca-tc 2017-05-11
@Gargron It seems that CPU usage has been optimized.

![924b536afd02fe5d94c274d2d526c7f1_720](https://cloud.githubusercontent.com/assets/1688137/25943805/18d88494-367c-11e7-8df6-56b154ab55ad.png)